refactor(itest): Move ApplicationConsole Matchers
They should be at the same location as other matchers.
