+++ itest/src/expect/Expectations.ts
@@ -1,16 +1,18 @@
 // noinspection JSUnusedGlobalSymbols
 
 // Import expect.extend
-import "./ElementHandleExpectations";
-import "./LocatorExpectations";
+import "./ApplicationConsoleExpectations";
 import "./ApplicationWrapperExpectations";
 import "./ClassicEditorWrapperExpectations";
+import "./ElementHandleExpectations";
+import "./LocatorExpectations";
 
 // Import Matcher Interfaces
-import { ElementHandleMatchers } from "./ElementHandleExpectations";
-import { LocatorMatchers } from "./LocatorExpectations";
 import { ApplicationWrapperMatchers } from "./ApplicationWrapperExpectations";
 import { ClassicEditorWrapperMatchers } from "./ClassicEditorWrapperExpectations";
+import { ElementHandleMatchers } from "./ElementHandleExpectations";
+import { LocatorMatchers } from "./LocatorExpectations";
+import { ApplicationConsoleMatchers } from "./ApplicationConsoleExpectations";
 
 /**
  * Tell TypeScript to know of new matchers.
@@ -19,20 +21,23 @@ declare global {
   // eslint-disable-next-line @typescript-eslint/no-namespace
   namespace jest {
     interface Expect
-      extends ApplicationWrapperMatchers,
+      extends ApplicationConsoleMatchers,
+        ApplicationWrapperMatchers,
         ClassicEditorWrapperMatchers,
         ElementHandleMatchers,
         LocatorMatchers {}
 
     // eslint-disable-next-line @typescript-eslint/ban-types
     interface Matchers<R = unknown, T = {}>
-      extends ApplicationWrapperMatchers<R, T>,
+      extends ApplicationConsoleMatchers<R, T>,
+        ApplicationWrapperMatchers<R, T>,
         ClassicEditorWrapperMatchers<R, T>,
         ElementHandleMatchers<R, T>,
         LocatorMatchers<R, T> {}
 
     interface InverseAsymmetricMatchers
-      extends ApplicationWrapperMatchers,
+      extends ApplicationConsoleMatchers,
+        ApplicationWrapperMatchers,
         ElementHandleMatchers,
         ClassicEditorWrapperMatchers,
         LocatorMatchers {}
